Extend GPS functionality and APRS message contents. Add locator calculation utility for WSPR/FT8. Fix reading of Si4032 temperature.

+++ src/codecs/ax25/ax25.c
@@ -23,7 +23,7 @@ static inline uint16_t ax25_calculate_crc_for_bit(uint16_t crc, bool bit)
 static inline uint16_t ax25_calculate_crc_for_byte(uint16_t crc, uint8_t byte)
 {
     uint8_t temp = byte;
-    for (int i = 0; i < 8; i++, temp >>= 1U) {
+    for (uint16_t i = 0; i < 8; i++, temp >>= 1U) {
         bool bit = ((temp & 1U) == 1U) ? true : false;
         crc = ax25_calculate_crc_for_bit(crc, bit);
     }
@@ -31,11 +31,11 @@ static inline uint16_t ax25_calculate_crc_for_byte(uint16_t crc, uint8_t byte)
     return crc;
 }
 
-static uint16_t ax25_calculate_crc(size_t length, uint8_t *data)
+static uint16_t ax25_calculate_crc(uint16_t length, uint8_t *data)
 {
     uint16_t crc = 0xFFFF;
 
-    for (size_t i = 0; i < length; i++) {
+    for (uint16_t i = 0; i < length; i++) {
         uint8_t byte = data[i];
         crc = ax25_calculate_crc_for_byte(crc, byte);
     }
@@ -43,19 +43,19 @@ static uint16_t ax25_calculate_crc(size_t length, uint8_t *data)
     return crc;
 }
 
-static size_t ax25_encode_digipeater_path(char *input, char *packet_data)
+static uint16_t ax25_encode_digipeater_path(char *input, char *packet_data)
 {
-    size_t digipeaters_length = (size_t) strlen(input);
-    size_t packet_data_index = 0;
+    uint16_t digipeaters_length = strlen(input);
+    uint16_t packet_data_index = 0;
 
-    for (size_t index = 0; index < digipeaters_length; index++) {
+    for (uint16_t index = 0; index < digipeaters_length; index++) {
         if (input[index] == ',' || index == digipeaters_length - 1) {
             if (input[index] != ',') {
                 packet_data[packet_data_index] = input[index] == '-' ? ' ' : input[index];
                 packet_data_index++;
             }
 
-            size_t fill_count = 7 - (packet_data_index % 7);
+            uint16_t fill_count = 7 - (packet_data_index % 7);
             while (fill_count > 0 && fill_count < 7) {
                 packet_data[packet_data_index] = ' ';
                 fill_count--;
@@ -72,8 +72,8 @@ static size_t ax25_encode_digipeater_path(char *input, char *packet_data)
     return packet_data_index;
 }
 
-size_t ax25_encode_packet_aprs(char *source, uint8_t source_ssid, char *destination, uint8_t destination_ssid,
-        char *digipeater_addresses, char *information_field, size_t length, uint8_t *packet_data)
+uint16_t ax25_encode_packet_aprs(char *source, uint8_t source_ssid, char *destination, uint8_t destination_ssid,
+        char *digipeater_addresses, char *information_field, uint16_t length, uint8_t *packet_data)
 {
     // TODO: use length to limit packet size
 
@@ -90,11 +90,11 @@ size_t ax25_encode_packet_aprs(char *source, uint8_t source_ssid, char *destinat
     header->destination_ssid = destination_ssid;
 
     char *digipeater_addresses_start = ((char *) header) + 1 + 14;
-    size_t digipeater_addresses_length = ax25_encode_digipeater_path(digipeater_addresses, digipeater_addresses_start);
+    uint16_t digipeater_addresses_length = ax25_encode_digipeater_path(digipeater_addresses, digipeater_addresses_start);
 
     // Perform bit-shifting for all addresses
     uint8_t *actual_data_start = ((uint8_t *) header) + 1;
-    for (int i = 0; i < 14 + digipeater_addresses_length; i++) {
+    for (uint16_t i = 0; i < 14 + digipeater_addresses_length; i++) {
         actual_data_start[i] = actual_data_start[i] << 1U;
     }
     actual_data_start[13 + digipeater_addresses_length] |= 1U;
@@ -105,10 +105,10 @@ size_t ax25_encode_packet_aprs(char *source, uint8_t source_ssid, char *destinat
     header_end->control_field = AX25_CONTROL_FIELD_UI_FRAME;
     header_end->protocol_id = AX25_PROTOCOL_ID_NO_LAYER_3;
 
-    size_t info_length = strlen(information_field);
+    uint16_t info_length = strlen(information_field);
     strcpy(header_end->information_field, information_field);
 
-    size_t crc_length = 14 + digipeater_addresses_length + 2 + info_length;
+    uint16_t crc_length = 14 + digipeater_addresses_length + 2 + info_length;
     uint16_t crc = ax25_calculate_crc(crc_length, actual_data_start);
 
     ax25_packet_footer *footer = (ax25_packet_footer *) (((uint8_t *) header_end->information_field) + info_length);
